requirements.txt
google-generativeai
PyMuPDF
python-docx
python-multipart
